refactor: remove route and place check logic in helper
