+++ api/controllers/EAppStartPageController.js
@@ -1,11 +1,10 @@
-const Application = require('../models/index').Application;
-const ApplicationReference = require('../models/index').ApplicationReference;
 
-const eAppServiceNum = 4;
 
 const EAppStartPageController = {
     startPage(req, res) {
         const userData = HelperService.getUserData(req, res);
+        const signInUrl = `${sails.config.customURLs.userServiceURL}/sign-in?from=start&next=continueEApp`;
+        const nextUrl = userData.loggedIn ? '/upload-files' : signInUrl;
         const backUrl =
             req.query.arrivedFrom === 'skipQuestionsLink'
                 ? '/eligibility/check-documents-are-eligible'
@@ -14,71 +13,8 @@ const EAppStartPageController = {
         return res.view('eApostilles/startPage.ejs', {
             user_data: userData,
             back_url: backUrl,
+            next_url: nextUrl,
         });
-    },
-
-    async startElectronicApplication(req, res) {
-        try {
-            const userData = HelperService.getUserData(req, res);
-            const signInUrl = `${sails.config.customURLs.userServiceURL}/sign-in?from=start&next=continueEApp`;
-            const nextUrl = userData.loggedIn ? '/upload-files' : signInUrl;
-
-            const appIdExistsInSession = Boolean(req.session.appId);
-            const appTypeExistsInSession = Boolean(req.session.appType);
-            const appDetailsInSession =
-                appIdExistsInSession && appTypeExistsInSession;
-
-            if (!appDetailsInSession) {
-                const newAppData =
-                    await EAppStartPageController._createNewApplicationInDB();
-                EAppStartPageController._addApplicationDetailsToSession(
-                    req,
-                    newAppData
-                );
-                EAppStartPageController._wipePreviousSessionVariables(req);
-            }
-
-            return res.redirect(nextUrl);
-        } catch (err) {
-            sails.log.error(err);
-            return res.serverError();
-        }
-    },
-
-    async _createNewApplicationInDB() {
-        try {
-            const appRefDetails = await ApplicationReference.findOne();
-            const uniqueApplicationId = HelperService.generateNewApplicationId(
-                appRefDetails,
-                eAppServiceNum.toString()
-            );
-            const newAppInDB = await Application.create({
-                serviceType: eAppServiceNum,
-                unique_app_id: uniqueApplicationId,
-                all_info_correct: '-1',
-                user_id: 0,
-                submitted: 'draft',
-                company_name: 'N/A',
-                feedback_consent: 0,
-                doc_reside_EU: 0,
-                residency: 0,
-            });
-
-            return newAppInDB;
-        } catch (err) {
-            throw new Error(`_createNewApplicationInDB: ${err}`);
-        }
-    },
-
-    _addApplicationDetailsToSession(req, newAppInDB) {
-        req.session.appId = newAppInDB.dataValues.application_id;
-        req.session.appType = eAppServiceNum;
-    },
-
-    _wipePreviousSessionVariables(req) {
-        req.session.selectedDocs = '';
-        req.session.return_address = '';
-        req.session.appSubmittedStatus = false;
     },
 };
