+++ views/eApostilles/startPage.ejs
@@ -118,7 +118,7 @@
         </ol>
       </div>
 
-      <a href="/checks-before-starting" role="button" draggable="false" class="govuk-button" data-module="govuk-button">
+      <a href="<%= next_url %>" role="button" draggable="false" class="govuk-button" data-module="govuk-button">
         Continue
       </a>
 </section>
